requirements.txt
python-twitter
typing

# Amazon AWS SDK
boto3